+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 - Adds (opt-out) experiment to disable cleaning up containers after a functions deploy (#6861)
 - Fix Next.js image optimization check in app directory for Windows (#6930)
 - Add support to next.config.mjs (#6933)
+- Remove development-only files from Next.js deployments (#6731)